license clarification and test apps CC zero
The test apps are intended to be adapted into user code freely,
this relicenses them public domain to make that clear.

Fixes and improvements are of course still welcome.

LICENSE clarifies now that some sources also have a more liberal license,
it was always told in the source itself but now it's summarized.

Signed-off-by: Andy Green <andy.green@linaro.org>