libuv: use shadow active events mask
libuv .events or .pevents in the io watcher don't seem to hold the POLLIN / POLLOUT
state correctly.  When POLLIN is defeated, using the rx flow control, POLLOUT seems
to go away somewhere along the line when trying to OR on to the active mask.

This has us track what we enabled outside in a container struct and use that for
the mask instead.  With this rx flow control works properly with the latest mirror
using libuv (it worked fine with poll() event loop without libuv).